requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
contextvars==2.4
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
google-auth==1.11.0
h11==0.9.0
h2==3.1.1
hpack==3.0.0
hstspreload==2020.2.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.8
immutables==0.11
oauthlib==3.1.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rfc3986==1.3.2
rsa==4.0
six==1.14.0
sniffio==1.1.0
sqlparse==0.3.0
urllib3==1.25.8
